Legger til async-storage for å ta vare på correlation ids
Rydder også i logger og legger på corrId/sessionId på alle log entries.
